+++ package.json
@@ -1,5 +1,6 @@
 {
   "name": "kekstagram",
+  "type": "module",
   "version": "30.0.0",
   "private": true,
   "description": "Личный проект «Кекстаграм» от HTML Academy",
@@ -24,4 +25,5 @@
     "node": "18",
     "npm": "9"
   }
+
 }
